Move data sync to common file.
